Read gh token from gh proper if not found in other locations (#107)
* Read gh token from gh proper if not found in other locations

GitHub CLI is moving away from storing the token in plain text configuration file. It will prefer to read the token from system's keyring storage.

As an unfortunate result, extensions will start failing now that token isn't in the YAML file anymore. But, we also don't want individual extensions to interface with the keyring. Instead, extensions will now shell out to `gh config get oauth_token --host HOSTNAME` to obtain the token.

This introduces a dependency on gh proper to be in PATH, but I consider that to be a reasonable request for extensions.

* Use gh auth token command

---------

Co-authored-by: Sam Coe <samcoe@users.noreply.github.com>