Replace duplicate isEnterprise and isTenancy fn in api pkg with auth pkg
There were duplicated, unexported isEnterprise and isTenancy functions in
the api package. The exported IsEnterprise and IsTenancy functions exist
in the auth package. This removes the duplicated functions from api and
replaces them with the exported packages from auth. It doesn't attempt to
clean up any more logic than that.